Add files via upload
The following repository contains the files required for designing a Noise Pollution monitoring system with the help of IOT. It uses firebase as a real-time server in which the data is uploaded with the help of a python script which tracks the readings from the prototype designed in Proteus Software with the help of a virtual port. A web app is deployed so as to make the user aware about the kind of data that is being uploaded. 
The repository includes the Arduino hex file, Python script, HTML CSS JS code, Proteus circuit software protoype and a report on the aforementioned project.